🏗️ Massive refactor around archive uncompression for reading/analysis
