+++ src/client/constants/action-types.ts
@@ -81,6 +81,8 @@ export const IMS_COMIC_BOOK_ARCHIVE_EXTRACTION_CALL_IN_PROGRESS =
 export const IMS_COMIC_BOOK_ARCHIVE_EXTRACTION_CALL_FAILED =
   "IMS_COMIC_BOOK_ARCHIVE_EXTRACTION_CALL_FAILED";
 
+export const COMICBOOK_EXTRACTION_SUCCESS = "COMICBOOK_EXTRACTION_SUCCESS";
+
 // Image file stats
 export const IMG_ANALYSIS_CALL_IN_PROGRESS = "IMG_ANALYSIS_CALL_IN_PROGRESS";
 export const IMG_ANALYSIS_DATA_FETCH_SUCCESS =
